Experimental WASI target for chdb
Adds a wasi-sdk build path alongside the existing Emscripten WASM target,
producing a plain .wasm that imports wasip1 syscalls with no JS glue.
wasm64-wasip1 is the primary target (matching the codebase's 64-bit size_t
assumptions); wasm32-wasip1 works as a fully-official-toolchain fallback.

Build system:
- cmake/wasi/toolchain.cmake + Platform/WASI.cmake for wasi-sdk integration
- cmake/target.cmake gains a WASI branch that reuses the whole OS_WASM
  feature trim and additionally defines OS_WASI, opts into wasi-libc's
  emulated signal/mman/process-clocks/getpid layers, and selects
  standardized exnref exception handling (-fwasm-exceptions with
  -wasm-use-legacy-eh=false, which wasmtime requires)
- CHDB_LITE stays forced on for WASM, but an explicit -DCHDB_LITE=OFF is
  now respected

Compat layer:
- cmake/wasi/compat-include/ supplies the POSIX surface wasi-libc lacks but
  the (Emscripten-shaped) OS_WASM paths expect: signals, sched, termios,
  pwd/grp, netdb, sockets that ENOSYS, ucontext that aborts, etc.
- programs/wasi/ entry point plus pthread mutex and misc libc stubs

Engine fixes, mostly 32-bit correctness and single-threaded degradation:
64-bit literals narrowed for 32-bit size_t/long, constexpr std::string
values exceeding 32-bit SSO capacity, std::variant integral ambiguity,
ProfileEvents::Count width, IAST_fwd stored_size, rusage field
availability, and OS_WASI guards around signal/timer machinery.

Status, build recipe, and remaining work are documented in
cmake/wasi/STATUS.md. Read queries (aggregation over numbers(), CSV via
file(), CREATE TABLE on the Log family) run under stock wasmtime;
MergeTree and boot-time table reloading still need a threads story.

+++ src/Common/threadPoolCallbackRunner.h
@@ -437,8 +437,8 @@ public:
     void shutdown();
 
 private:
-    static constexpr Int64 SHUTDOWN_START = 1l << 42; // shutdown requested
-    static constexpr Int64 SHUTDOWN_END = 1l << 52; // no shared locks remain
+    static constexpr Int64 SHUTDOWN_START = Int64(1) << 42; // shutdown requested (1l overflows on 32-bit targets)
+    static constexpr Int64 SHUTDOWN_END = Int64(1) << 52; // no shared locks remain
 
     /// If >= SHUTDOWN_START, no new try_lock_shared() calls will succeed.
     /// Whoever changes the value to exactly SHUTDOWN_START (i.e. shutdown requested, no shared locks)
